% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/val_input.R
\name{val_input}
\alias{val_input}
\title{Check variable inputs}
\usage{
val_input(input, input_type, ...)
}
\arguments{
\item{input}{An argument.}

\item{input_type}{The expected type of the argument. (Currently supported: \code{"character"}, \code{"numeric"}, \code{"logical"})}

\item{input_validation}{Optional.}

\item{min_value}{.}

\item{max_value}{.}

\item{max_length}{.}

\item{variable_name}{.}

\item{is_missing}{.}
}
\value{
nothing
}
\description{
Internal function to check input variables
}
